test(architecture): give the repository-wide guards their own home
Three source-tree guards were living in the suite for one service helper:
the ban on single-row deletes of chat message rows, the sweep that stops
a message type from being written back against its one-way direction, and
the layering rule that keeps the chat history service independent of the
interaction rollout controls. None of the three is a fact about that
helper -- each constrains the whole tree, or the dependency graph, and
each would still hold if that helper were deleted tomorrow. Filed under a
feature, they were findable only by someone who already knew where to
look, and the next repository-wide rule had no obvious place to go.

They move, with the shared source-parsing fixture and every positive and
negative control, into a module whose docstring states what earns a place
there. The mid-turn persistence guard stays behind: it pins a value that
belongs to that one feature. It keeps using the constant-resolution
helpers, now imported from the new module.

Pure move: no guard's body, docstring, assertion message or fixture table
changes, and the test count is unchanged on both sides. Review with
`git show -M -C` -- the diff is large and the reviewable surface is the
new module docstring plus the import lines.
